+++ docs/changelog.md
@@ -10,7 +10,15 @@ by release.
 - milestone: [1.17.0](https://github.com/enricoros/big-agi/milestone/17)
 - work in progress: [big-AGI open roadmap](https://github.com/users/enricoros/projects/4/views/2), [help here](https://github.com/users/enricoros/projects/4/views/4)
 
-### What's New in 1.16.1 · May 13, 2024 (minor release, models support)
+### What's New in 1.16.2 · Jun 7, 2024 (minor release)
+
+- Improve web downloads, as text, markdwon, or HTML
+- Proper support for Gemini models
+- Added the latest Mistral model
+- Tokenizer support for gpt-4o
+- Updates to Beam
+
+### What's New in 1.16.1 · May 13, 2024 (minor release)
 
 - Support for the new OpenAI GPT-4o 2024-05-13 model